backend/requirements.txt
faiss-cpu==1.9.0
rank-bm25==0.2.2
groq==1.1.2
pymupdf==1.24.14
langchain
langchain-core
python-dotenv==1.0.1
numpy==1.26.4
Pillow==10.4.0
pytesseract==0.3.13
sentence-transformers==2.7.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0